+++ libifstate/util.py
@@ -13,8 +13,25 @@ import socket
 import fcntl
 import struct
 import array
+import struct
 
 # ethtool helper
+ETHTOOL_GDRVINFO = 0x00000003  # Get driver info
+STRUCT_DRVINFO = struct.Struct(
+    "I" +    # cmd
+    "32s" +  # driver
+    "32s" +  # version
+    "32s" +  # fw_version
+    "32s" +  # bus_info
+    "32s" +  # reserved1
+    "12s" +  # reserved2
+    "I" +    # n_priv_flags
+    "I" +    # n_stats
+    "I" +    # testinfo_len
+    "I" +    # eedump_len
+    "I"      # regdump_len
+)
+
 ETHTOOL_GPERMADDR = 0x00000020  # Get permanent hardware address
 L2_ADDRLENGTH = 6  # L2 address length
 sock = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
@@ -36,6 +53,23 @@ class IPRouteExt(IPRoute):
             NLM_F_ACK | NLM_F_CREATE | NLM_F_EXCL
         ))
 
+    def get_businfo(self, ifname):
+        data = array.array("B", struct.pack(
+            "I", ETHTOOL_GDRVINFO))
+        data.extend(b'\x00' * (STRUCT_DRVINFO.size - len(data)))
+
+        ifr = struct.pack('16sP', ifname.encode(
+            "utf-8"), data.buffer_info()[0])
+
+        try:
+            r = fcntl.ioctl(sock.fileno(), SIOCETHTOOL, ifr)
+        except OSError:
+            return None
+
+        drvinfo = STRUCT_DRVINFO.unpack(data)
+
+        return drvinfo[4].decode('ascii').split('\x00')[0]
+
     def get_permaddr(self, ifname):
         data = array.array("B", struct.pack(
             "II", ETHTOOL_GPERMADDR, L2_ADDRLENGTH))
@@ -54,6 +88,14 @@ class IPRouteExt(IPRoute):
             return None
 
         return l2addr
+
+    def get_iface_by_businfo(self, businfo):
+        for iface in iter(self.get_links()):
+            ifname = iface.get_attr('IFLA_IFNAME')
+            bi = self.get_businfo(ifname)
+
+            if bi and bi == businfo:
+                return iface['index']
 
     def get_iface_by_permaddr(self, permaddr):
         for iface in iter(self.get_links()):